SeedRestrictions
Created hashing function to easier identify each restriction list.
Created restriction type and altered the code to comply with accepting and returning restriction type values which are written used snake_case as to be compatible with the supabase table schema.
Created parseRestrictions function which return all the so far designed restrictions.