Seed restrictions
Moved the funcitonality that gets all the restrictions to the utils file.
Created functionality that checks for the restrictions existence in supabase, before trying to insert it. If the restrictions exists, the code then compares the hash value of the already existing restriction list to the one that is stored on the app. If the hash values are equal it means the list is up to date. If not then the restrictions champion_list and hash values are updated.
Created deno task seedRestrictions